registry/test-plugin/pyproject.toml
[project]
name = "hivemind-plugin-demo"
version = "0.9.0"
description = "Minimal hivemind plugin for testing the registry (one demo tool)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "hivemind-ai",
]

[project.entry-points."hivemind.plugins"]
demo = "hivemind_plugin_demo:load"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hivemind_plugin_demo"]
